Moved AMS_5600_PI.hpp and updated references.
The AMS_5600_PI.hpp header was in a sub-directory named for the
Adafruit mux (TCA9548). This directory contained nothing to do
with the mux which doesn't have an associated library, and also
contained no other files so I moved the header to include/bb_sensors
with the others and updated the relevant references.

Tested by compiling the codebase but further testing is required
to verify functionality.

Removed