Updated the name of the i2c_mux source and executable to wind_mux.
Changes:
Updated the name of i2c_mux to wind_mux to be more appropriate.
I also updated launch files and the relevant documentation.

Testing:
Checked everything still compiled and ran as before.

Note:
I did not re-build the documentation as this needs to be done
on the host machine.